+++ src/main/kotlin/com/github/gplassard/managedexcludes/services/config/ManagedExcludesConfigService.kt
@@ -1,7 +1,7 @@
-package com.github.gplassard.managedexcludes.services
+package com.github.gplassard.managedexcludes.services.config
 
 import com.github.gplassard.managedexcludes.Constants
-import com.intellij.openapi.application.ReadAction
+import com.intellij.openapi.application.readAction
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.fileEditor.FileDocumentManager
@@ -11,22 +11,21 @@ import com.intellij.psi.search.FilenameIndex
 import com.intellij.psi.search.GlobalSearchScope
 
 @Service(Service.Level.PROJECT)
-class ConfigService {
-
-    fun loadExcludeConfig(project: Project): Set<VirtualFile> {
-        val excludeFiles = ReadAction.compute<Collection<VirtualFile>, RuntimeException> {
-            FilenameIndex.getVirtualFilesByName(
+class ManagedExcludesConfigService {
+    suspend fun loadExcludeConfig(project: Project): Set<VirtualFile> {
+         return readAction {
+             val excludeFiles = FilenameIndex.getVirtualFilesByName(
                 Constants.EXCLUDE_FILE_NAME,
                 true,
                 GlobalSearchScope.projectScope(project)
             )
-        }
-        thisLogger().info("Excluding files found $excludeFiles")
+             thisLogger().info("Excluding files found $excludeFiles")
 
-        return excludeFiles
-            .flatMap { resolveRelativeFiles(it) }
-            .toSet()
-            .also { thisLogger().info("Aggregation done, there is a total of ${it.size} distinct paths to exclude") }
+             excludeFiles
+                 .flatMap { resolveRelativeFiles(it) }
+                 .toSet()
+                 .also { thisLogger().info("Aggregation done, there is a total of ${it.size} distinct paths to exclude") }
+        }
     }
 
     private fun resolveRelativeFiles(excludeFile: VirtualFile): List<VirtualFile> =
@@ -43,5 +42,4 @@ class ConfigService {
             ?.filter { it.exists() }
             ?.toList()
             .orEmpty()
-
 }